requirements.txt
anyio==4.9.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.0
crashtest==0.4.1
cryptography==45.0.1
distlib==0.3.9
dulwich==0.22.8
exceptiongroup==1.3.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
fugashi==1.4.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
keyring==25.6.0
mecab==0.996.11
more-itertools==10.7.0
msgpack==1.1.0
packaging==25.0
pbs-installer==2025.4.9
pkginfo==1.12.1.2
plac==1.4.5
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
pycparser==2.22
pyproject_hooks==1.2.0
python-dotenv==1.1.0
RapidFuzz==3.13.0
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
strip-tags==0.6
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.5.9.12
typing_extensions==4.13.2
unidic==1.1.0
unidic-lite==1.0.8
urllib3==2.4.0
virtualenv==20.31.2
wasabi==0.10.1
webencodings==0.5.1
zipp==3.21.0
zstandard==0.23.0
